+++ actions/users/delete.php
@@ -2,7 +2,7 @@
 include_once('../../config/init.php');
 include_once($BASE_DIR .'database/users.php');
 
-
+checkIfAdmin();
 deleteUserId($_POST['id']);
 
 header('Location: '.$BASE_URL.'pages/users/view_users.php');
